# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Ride-my-way is an online platform where users can order for rides to their individual destination",
    "version": "1.0.0",
    "title": "Ride-my-way",
    "contact": {
      "email": "bosomadex@gmail.com"
    }
  },
  "host": "localhost:2021",
  "basePath": "/api",
...
  "schemes": ["http", "https"],
  "paths": {
    "/user/create": {
      "post": {
        "tags": ["Users"],
        "summary": "Create user account",
        "description": "This will create user account",
        "operationId": "createUserAccount",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that need to be added to user table",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success Response Format",
            "schema": {
              "$ref": "#/definitions/SuccessApiResponseFormat"
            }
          },
          "400": {
            "description": "Failure Response Format",
            "schema": {
              "$ref": "#/definitions/FailedApiResponseFormat"
            }
          }
        }
      }
    },
...
    "/user/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login user",
        "description": "Allow users to login to the platform",
        "operationId": "loginUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that need to be logged in",
            "required": true,
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success Response Format",
            "schema": {
              "$ref": "#/definitions/SuccessApiResponseFormat"
            }
          },
          "400": {
            "description": "Failure Response Format",
            "schema": {
              "$ref": "#/definitions/FailedApiResponseFormat"
            }
          }
        }
      }
    },
    "/ride": {
      "get": {
        "tags": ["Ride"],
        "summary": "get all ride",
        "description": "Multiple Ride to be displayed",
        "operationId": "getAllRides",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "$ref": "#/definitions/token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success Response Format",
            "schema": {
              "$ref": "#/definitions/SuccessApiResponseFormat"
            }
          },
          "400": {
            "description": "Failure Response Format",
            "schema": {
              "$ref": "#/definitions/FailedApiResponseFormat"
            }
          }
        }
      }
    },
    "/ride/create": {
      "post": {
        "tag": ["Ride"],
        "summary": "create ride for users",
        "description": "Allow user to create ride",
        "operationId": "createRideOffer",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "Bearer token",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/token"
            }
          },
          {
            "in": "body",
            "name": "body",
            "type": "object",
            "description": "Ride list",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ride"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success Response Format"
          },
          "400": {
            "description": "Failure Response Format"
          }
        }
      }
    },
    "/ride/{rideId}": {
      "get": {
        "tags": ["Ride"],
        "summary": "get One Ride",
        "description": "One Ride to be displayed",
        "operationId": "getOneRide",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "Bearer token",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/token"
            }
          },
          {
            "in": "path",
            "name": "rideId",
            "description": "params",
            "required": true,
            "schema": {
              "$ref": "#/definitions/params"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success Response Format",
            "schema": {
              "$ref": "#/definitions/SuccessApiResponseFormat"
            }
          },
          "400": {
            "description": "Failure Response Format",
            "schema": {
              "$ref": "#/definitions/FailedApiResponseFormat"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "params": {
      "types": "string",
      "example": "5ee4ef511bd45f1010f12ca4"
    },
    "token": {
      "type": "string",
      "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVlZTRlZjUxMWJkNDVmMTAxMGYxMmNhNCIsImlhdCI6MTU5MjA2MjgxMSwiZXhwIjoxNTkyMTQ5MjExfQ.ZbOs13ffTE0xmUFM8fXH70lDrTMzodm1KAYEwwq-rL0"
    },
    "ride": {
      "type": "object",
      "properties": {
        "departure": {
          "type": "string",
          "example": "example"
        },
        "destination": {
          "type": "string",
          "example": "example"
        },
        "seats": {
          "type": "string",
          "example": "example"
        },
        "time": {
          "type": "string",
          "example": "example"
        },
        "date": {
          "type": "string",
          "example": "example"
        }
      }
    },
...
    "user": {
      "type": "object",
      "properties": {
        "fullName": {
          "type": "string",
          "example": "fullName"
        },
        "username": {
          "type": "string",
          "example": "username"
        },
        "phoneNumber": {
          "type": "string",
          "example": "phoneNumber"
        },
        "email": {
          "type": "string",
          "example": "email"
        },
        "password": {
          "type": "string",
          "format": "password",
          "example": "password"
        }
      }
    },
    "login": {
      "type": "object",
      "properties": {
        "dataField": {
          "type": "string",
          "example": "bosomadex@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "password"
        }
      }
    },
    "SuccessApiResponseFormat": {
      "type":"object",
      "properties":{
        "message": {
          "type":"string"
        },
        "data": {
          "type":"object"
        }
      }
    },
    "FailedApiResponseFormat": {
      "type":"object",
      "properties":{
        "message": {
          "type":"string"
        },
        "error": {
          "type":"object"
        }
      }
    }
  },
...
}